docs: ✏️ add example applications
