--- examples/azure-function/host.json
@@ -1,12 +1,0 @@
-{
-  "version": "2.0",
-  "extensionBundle": {
-    "id": "Microsoft.Azure.Functions.ExtensionBundle",
-    "version": "[1.*, 2.0.0)"
-  },
-  "logging": {
-    "console": {
-      "isEnabled": "true"
-    }
-  }
-}
