--- examples/azure-function/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "generated-functions-api",
-  "version": "1.0.0",
-  "scripts": {
-    "build": "tsc",
-    "watch": "tsc -w",
-    "prestart": "npm run build",
-    "start": "func start"
-  },
-  "dependencies": {
-    "azure-functions-ts-essentials": "1.3.2",
-    "dotenv": "8.2.0",
-    "knex": "0.21.2",
-    "mssql": "6.2.1",
-    "express": "^4.17.1",
-    "@inductjs/core": "0.2.5"
-  },
-  "devDependencies": {
-    "@azure/functions": "^1.2.0",
-    "@types/knex": "0.16.1",
-    "@types/express": "^4.17.7",
-    "@types/mssql": "6.0.0",
-    "@types/uuid": "7.0.2",
-    "typescript": "3.8.3"
-  },
-  "author": "Induct API Generator",
-  "license": "UNLICENSED",
-  "homepage": "https://github.com/Yeseh/induct-generator"
-}
