+++ examples/express-app-mongo/models/user-model.ts
@@ -1,0 +1,15 @@
+import { prop } from "@inductjs/core";
+
+class UserSchema {
+    @prop()
+    public id?: number
+
+    @prop()
+    public userName?: string;
+
+    constructor(args: UserSchema) {
+        Object.assign(this, args);
+    }
+}
+
+export {UserSchema};
