+++ examples/express-app-sql/src/index.ts
@@ -1,0 +1,9 @@
+require("dotenv").config();
+import {createServer} from "@inductjs/core";
+
+(async function() {
+  // InductServer auto mounts routers from src/routers ending in -router.{js, ts}
+  const server = await createServer();
+
+  server.start();
+})();
