+++ package.json
@@ -45,16 +45,15 @@
         }
     },
     "devDependencies": {
+        "@inductjs/core": "^0.5.4",
         "@types/express": "^4.17.7",
         "@types/jest": "^26.0.13",
         "@types/knex": "^0.16.1",
-        "@types/mongoose": "^5.7.36",
         "@types/node": "^14.0.1",
         "@types/rimraf": "^3.0.0",
         "@typescript-eslint/parser": "^2.33.0",
         "@yeseh/eslint-config": "git://github.com/yeseh/eslint-config.git#master",
         "@yeseh/prettier-config": "git://github.com/yeseh/prettier-config.git#master",
-        "chalk": "^4.1.0",
         "eslint": "^7.5.0",
         "git-cz": "^4.7.1",
         "jest": "^26.4.2",
@@ -74,12 +73,14 @@
     "dependencies": {
         "@azure/functions": "^1.2.2",
         "@typegoose/typegoose": "^7.4.1",
+        "@types/mongoose": "^5.7.36",
         "azure-functions-ts-essentials": "^1.3.2",
         "body-parser": "^1.19.0",
         "class-validator": "^0.12.2",
         "express": "^4.17.1",
         "knex": "^0.21.2",
         "mongoose": "^5.10.8",
-        "ts-transformer-keys": "^0.4.2"
+        "ts-transformer-keys": "^0.4.2",
+        "chalk": "^4.1.0"
     }
 }
